requirements.txt
boto3==1.42.8
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-genai==1.56.0
livekit==1.0.20
livekit-agents==1.3.9
livekit-api==1.1.0
livekit-plugins-deepgram==1.3.6
livekit-plugins-google==1.3.9
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.3.6
livekit-plugins-silero==1.3.6
livekit-plugins-turn-detector==1.3.6
livekit-protocol==1.1.1
openai==2.11.0
pyodbc==5.3.0
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
tzdata